WebVacations/requirements/dev.txt
Django==2.2.10
django-mssql-backend==2.8.1
Pillow==9.3.0
django-crispy-forms==1.14.0
reportlab==3.6.12